Build groupBy groups imperatively
